+++ src/main/java/lotto/domain/input/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.input;
 
 import lotto.view.InputView;
