+++ src/main/java/lotto/domain/prize/PrizeMoney.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.prize;
 
 public enum PrizeMoney {
