+++ src/main/java/lotto/domain/prize/PrizeSystem.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.prize;
+
+import lotto.domain.lotto.Lotto;
 
 import java.util.List;
